# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GlobalHR Platform API - HR, Payroll, Contracts, Compliance, Onboarding
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List organizations
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get an organization by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found

  # ─── WORKERS ──────────────────────────────────────────────────────────────
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, pending_signature, active, expired, terminated]
        - name: workerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/sign:
    post:
      operationId: signContract
      tags: [contracts]
      summary: Sign a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signed contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"

  # ─── PAYMENTS ─────────────────────────────────────────────────────────────
...
  /compliance:
    get:
      operationId: listComplianceItems
      tags: [compliance]
      summary: List compliance items
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_review, approved, rejected, expired]
        - name: workerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of compliance items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceItem"
    post:
      operationId: createComplianceItem
      tags: [compliance]
      summary: Create a compliance document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplianceItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
  /compliance/{id}:
    get:
      operationId: getComplianceItem
      tags: [compliance]
      summary: Get a compliance item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Compliance item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
        "404":
          description: Not found
    patch:
      operationId: updateComplianceItem
      tags: [compliance]
      summary: Update compliance status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplianceItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"

  # ─── ONBOARDING ───────────────────────────────────────────────────────────
  /onboarding:
    get:
      operationId: listOnboardingTasks
      tags: [onboarding]
      summary: List onboarding tasks
      parameters:
        - name: workerId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed, skipped]
      responses:
        "200":
          description: List of onboarding tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnboardingTask"
    post:
      operationId: createOnboardingTask
      tags: [onboarding]
      summary: Create an onboarding task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOnboardingTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingTask"
  /onboarding/{id}:
    patch:
      operationId: updateOnboardingTask
      tags: [onboarding]
      summary: Update onboarding task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOnboardingTaskBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingTask"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        legalName:
          type: ["string", "null"]
        country:
          type: string
        industry:
          type: string
        size:
          type: string
          enum: [startup, smb, enterprise]
        logoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, country, industry, size, createdAt, updatedAt]
    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        legalName:
          type: ["string", "null"]
        country:
          type: string
        industry:
          type: string
        size:
          type: string
          enum: [startup, smb, enterprise]
        logoUrl:
          type: ["string", "null"]
      required: [name, country, industry, size]

    # ─── WORKER ───────────────────────────────────────────────────────────────
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        organizationId:
          type: integer
        title:
          type: string
        contractType:
          type: string
          enum: [full_time, part_time, fixed_term, indefinite, freelance]
        status:
          type: string
          enum: [draft, pending_signature, active, expired, terminated]
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        currency:
          type: string
        compensation:
          type: number
        compensationPeriod:
          type: string
          enum: [hourly, monthly, annual]
        signedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, workerId, organizationId, title, contractType, status, startDate, currency, compensation, compensationPeriod, createdAt, updatedAt]
    CreateContractBody:
      type: object
      properties:
        workerId:
          type: integer
        organizationId:
          type: integer
        title:
          type: string
        contractType:
          type: string
          enum: [full_time, part_time, fixed_term, indefinite, freelance]
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
        currency:
          type: string
        compensation:
          type: number
        compensationPeriod:
          type: string
          enum: [hourly, monthly, annual]
      required: [workerId, organizationId, title, contractType, startDate, currency, compensation, compensationPeriod]
    UpdateContractBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [draft, pending_signature, active, expired, terminated]
        endDate:
          type: ["string", "null"]
        compensation:
          type: number

    # ─── PAYMENT ─────────────────────────────────────────────────────────────
...
    ComplianceItem:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        organizationId:
          type: integer
        documentType:
          type: string
          enum: [work_permit, tax_id, id_document, right_to_work, background_check, other]
        status:
          type: string
          enum: [pending, in_review, approved, rejected, expired]
        country:
          type: string
        notes:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, workerId, organizationId, documentType, status, country, createdAt, updatedAt]
    CreateComplianceItemBody:
      type: object
      properties:
        workerId:
          type: integer
        organizationId:
          type: integer
        documentType:
          type: string
          enum: [work_permit, tax_id, id_document, right_to_work, background_check, other]
        country:
          type: string
        notes:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
      required: [workerId, organizationId, documentType, country]
    UpdateComplianceItemBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_review, approved, rejected, expired]
        notes:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]

    # ─── ONBOARDING ───────────────────────────────────────────────────────────
    OnboardingTask:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        organizationId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [documents, equipment, access, training, compliance, introduction, other]
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        dueDate:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, workerId, organizationId, title, category, status, createdAt, updatedAt]
    CreateOnboardingTaskBody:
      type: object
      properties:
        workerId:
          type: integer
        organizationId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [documents, equipment, access, training, compliance, introduction, other]
        dueDate:
          type: ["string", "null"]
      required: [workerId, organizationId, title, category]
    UpdateOnboardingTaskBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        description:
          type: ["string", "null"]

    # ─── NOTIFICATION ─────────────────────────────────────────────────────────
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [contract, payment, compliance, onboarding, general]
        isRead:
          type: boolean
        workerId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, title, message, type, isRead, createdAt]

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...