# lib/api-zod/src/generated/types/createContractBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlobalHR Platform API - HR, Payroll, Contracts, Compliance, Onboarding
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContractBodyCompensationPeriod } from "./createContractBodyCompensationPeriod";
import type { CreateContractBodyContractType } from "./createContractBodyContractType";

export interface CreateContractBody {
  workerId: number;
  organizationId: number;
  title: string;
  contractType: CreateContractBodyContractType;
  startDate: Date;
  /** @nullable */
  endDate?: string | null;
  currency: string;
  compensation: number;
  compensationPeriod: CreateContractBodyCompensationPeriod;
}