# lib/api-zod/src/generated/types/createOrganizationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlobalHR Platform API - HR, Payroll, Contracts, Compliance, Onboarding
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrganizationBodySize } from "./createOrganizationBodySize";

export interface CreateOrganizationBody {
  name: string;
  /** @nullable */
  legalName?: string | null;
  country: string;
  industry: string;
  size: CreateOrganizationBodySize;
  /** @nullable */
  logoUrl?: string | null;
}